Refactor asset provider to handle JS asset setup
