+++ plugin/src/software/aws/toolkits/eclipse/amazonq/views/AmazonQChatWebview.java
@@ -3,26 +3,13 @@
 
 package software.aws.toolkits.eclipse.amazonq.views;
 
-import java.awt.Toolkit;
-import java.awt.datatransfer.StringSelection;
-import java.util.Optional;
-
 import org.eclipse.swt.browser.Browser;
-import org.eclipse.swt.browser.BrowserFunction;
-import org.eclipse.swt.browser.ProgressAdapter;
-import org.eclipse.swt.browser.ProgressEvent;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Display;
 
 import software.aws.toolkits.eclipse.amazonq.chat.ChatCommunicationManager;
 import software.aws.toolkits.eclipse.amazonq.chat.ChatStateManager;
-import software.aws.toolkits.eclipse.amazonq.chat.ChatTheme;
-import software.aws.toolkits.eclipse.amazonq.plugin.Activator;
 import software.aws.toolkits.eclipse.amazonq.providers.assets.ChatWebViewAssetProvider;
 import software.aws.toolkits.eclipse.amazonq.providers.assets.WebViewAssetProvider;
-import software.aws.toolkits.eclipse.amazonq.util.PluginPlatform;
-import software.aws.toolkits.eclipse.amazonq.util.PluginUtils;
-import software.aws.toolkits.eclipse.amazonq.util.ThreadingUtils;
 import software.aws.toolkits.eclipse.amazonq.views.actions.AmazonQCommonActions;
 
 public class AmazonQChatWebview extends AmazonQView implements ChatUiRequestListener {
@@ -31,24 +18,16 @@ public class AmazonQChatWebview extends AmazonQView implements ChatUiRequestList
 
     private AmazonQCommonActions amazonQCommonActions;
     private final ChatStateManager chatStateManager;
-    private final ViewCommandParser commandParser;
-    private final ViewActionHandler actionHandler;
     private final ChatCommunicationManager chatCommunicationManager;
-    private final ChatTheme chatTheme;
     private Browser browser;
     private volatile boolean canDisposeState = false;
     private WebViewAssetProvider webViewAssetProvider;
-    private Optional<String> content;
 
     public AmazonQChatWebview() {
         super();
         this.chatStateManager = ChatStateManager.getInstance();
-        this.commandParser = new LoginViewCommandParser();
         this.chatCommunicationManager = ChatCommunicationManager.getInstance();
-        this.actionHandler = new AmazonQChatViewActionHandler(chatCommunicationManager);
         this.webViewAssetProvider = new ChatWebViewAssetProvider();
-        this.content = this.webViewAssetProvider.getContent();
-        this.chatTheme = new ChatTheme();
     }
 
     @Override
@@ -67,23 +46,7 @@ public class AmazonQChatWebview extends AmazonQView implements ChatUiRequestList
             }
 
             browser = getAndUpdateStateManager();
-
-            browser.setVisible(false);
-            browser.addProgressListener(new ProgressAdapter() {
-                @Override
-                public void completed(final ProgressEvent event) {
-                    Display.getDefault().asyncExec(() -> {
-                        if (!browser.isDisposed()) {
-                            browser.setVisible(true);
-                        }
-                    });
-                }
-            });
-
-            if (!content.isPresent()) {
-                content = webViewAssetProvider.getContent();
-            }
-            browser.setText(content.get());
+            webViewAssetProvider.injectAssets(browser);
         } else {
             updateBrowser(browser);
         }
@@ -95,53 +58,8 @@ public class AmazonQChatWebview extends AmazonQView implements ChatUiRequestList
 
         chatCommunicationManager.setChatUiRequestListener(this);
 
-        new BrowserFunction(browser, "ideCommand") {
-            @Override
-            public Object function(final Object[] arguments) {
-                ThreadingUtils.executeAsyncTask(() -> {
-                    handleMessageFromUI(browser, arguments);
-                });
-                return null;
-            }
-        };
 
         addFocusListener(parent, browser);
-
-        new BrowserFunction(browser, "isMacOs") {
-            @Override
-            public Object function(final Object[] arguments) {
-                return Boolean.TRUE.equals(PluginUtils.getPlatform() == PluginPlatform.MAC);
-            }
-        };
-
-        new BrowserFunction(browser, "copyToClipboard") {
-            @Override
-            public Object function(final Object[] arguments) {
-                if (arguments.length > 0 && arguments[0] instanceof String) {
-                    StringSelection stringSelection = new StringSelection((String) arguments[0]);
-                    Toolkit.getDefaultToolkit().getSystemClipboard().setContents(stringSelection, null);
-                }
-                return null;
-            }
-        };
-
-        addFocusListener(parent, browser);
-
-        // Inject chat theme after mynah-ui has loaded
-        browser.addProgressListener(new ProgressAdapter() {
-            @Override
-            public void completed(final ProgressEvent event) {
-                Display.getDefault().syncExec(() -> {
-                    try {
-                        chatTheme.injectTheme(browser);
-                        disableBrowserContextMenu();
-                    } catch (Exception e) {
-                        Activator.getLogger().info("Error occurred while injecting theme into Q chat", e);
-                    }
-                });
-            }
-        });
-
         setupAmazonQCommonActions();
 
         return parent;
@@ -151,15 +69,6 @@ public class AmazonQChatWebview extends AmazonQView implements ChatUiRequestList
         var browser = getBrowser();
         chatStateManager.updateBrowser(browser);
         return browser;
-    }
-
-    private void handleMessageFromUI(final Browser browser, final Object[] arguments) {
-        try {
-            commandParser.parseCommand(arguments)
-                    .ifPresent(parsedCommand -> actionHandler.handleCommand(parsedCommand, browser));
-        } catch (Exception e) {
-            Activator.getLogger().error("Error processing message from Amazon Q chat", e);
-        }
     }
 
     @Override
@@ -180,9 +89,7 @@ public class AmazonQChatWebview extends AmazonQView implements ChatUiRequestList
         if (canDisposeState) {
             ChatStateManager.getInstance().dispose();
             webViewAssetProvider.dispose();
-            content = Optional.empty();
         }
-        webViewAssetProvider.dispose();
         super.dispose();
     }
 }
